Dunno why but still on some songs the score counter shows negative score. trying to fix

+++ Renderers/UIElements/ScoreMeter.h
@@ -15,11 +15,11 @@
 @class FontString, Texture2D;
 
 @interface ScoreMeter : NSObject <TMLogicUpdater, TMRenderable, TMMessageSupport> {
-	int				m_nCurrentScore;
+	long			m_nCurrentScore;
 	int				m_nTotalSteps;
 	int				m_nTapNotesHit;
 	int				m_nDifficulty;
-	int				m_nMaxPossiblePoints;
+	long			m_nMaxPossiblePoints;
 	int				m_nRoundTo;
 	int				m_nScoreRemainder;
